refactor mock stripe into modules for invoice.created prep
